requirements.txt
Eve==0.5.3
gunicorn==19.3.0
Flask-Script==2.0.5
PyJWT
